// Source/Client/Interop/JSInteropTypes.Generated.ts
export interface IFileInterop {
    appendContents(filePath: string, contents: string): Promise<string>;
    copyFile(sourceFilePath: string, destinationFilePath: string): Promise<string>;
    createDirectory(directoryPath: string): Promise<string>;
    deleteDirectory(directoryPath: string): Promise<string>;
    deleteFile(filePath: string): Promise<string>;
    getDirectories(directoryPath: string): Promise<string[]>;
    getFiles(directoryPath: string): Promise<string[]>;
    getTemporaryFilePath(): Promise<string>;
    insertLine(filePath: string, lineNumber: number, contents: string): Promise<string>;
    readContents(filePath: string): Promise<string>;
    readLine(filePath: string, lineNumber: number): Promise<string>;
    writeContents(filePath: string, contents: string): Promise<string>;
    writeLine(filePath: string, lineNumber: number, contents: string): Promise<string>;
}